moved the list to the Autocomplete component and polished query properties
