+++ src/data/user.ts
@@ -1,0 +1,9 @@
+export type User = {
+  name: string;
+  sex: string;
+  born: number;
+  died: number;
+  fatherName: string;
+  motherName: string;
+  slug: string;
+};
